# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DropParty API - Viral onchain USDC giveaway platform on Arc Network
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Drop {
  id: number;
  contractAddress: string;
  creatorAddress: string;
  title: string;
  /** Total USDC amount in the drop (as string to handle large numbers) */
  totalAmount: string;
  /** Amount each claimer receives in USDC */
  amountPerClaim: string;
  maxClaims: number;
  claimedCount: number;
  isActive: boolean;
  /**
   * ISO 8601 timestamp when the drop expires
   * @nullable
   */
  expiresAt: string | null;
  /** ISO 8601 timestamp */
  createdAt: string;
  /**
   * Transaction hash of the drop creation
   * @nullable
   */
  txHash: string | null;
}

export interface CreateDropBody {
  contractAddress: string;
  creatorAddress: string;
  title: string;
  totalAmount: string;
  amountPerClaim: string;
  maxClaims: number;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  txHash?: string | null;
}

export interface Claim {
  id: number;
  dropId: number;
  claimerAddress: string;
  amount: string;
  /** @nullable */
  txHash: string | null;
  /** ISO 8601 timestamp */
  claimedAt: string;
}

export interface RecordClaimBody {
  claimerAddress: string;
  amount: string;
  /** @nullable */
  txHash?: string | null;
}

export interface ActivityItem {
  claimerAddress: string;
  amount: string;
  dropTitle: string;
  contractAddress: string;
  claimedAt: string;
}

export interface DropsSummary {
  /** Total USDC ever dropped */
  totalDropped: string;
  totalClaims: number;
  activeDropsCount: number;
  totalDropsCount: number;
}

export interface ErrorResponse {
  error: string;
}