# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DropParty API - Viral onchain USDC giveaway platform on Arc Network
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns a list of all drops with their current status
 * @summary List all drops
 */
export const ListDropsResponseItem = zod.object({
  id: zod.number(),
  contractAddress: zod.string(),
  creatorAddress: zod.string(),
  title: zod.string(),
  totalAmount: zod
    .string()
    .describe(
      "Total USDC amount in the drop (as string to handle large numbers)",
    ),
  amountPerClaim: zod.string().describe("Amount each claimer receives in USDC"),
  maxClaims: zod.number(),
  claimedCount: zod.number(),
  isActive: zod.boolean(),
  expiresAt: zod
    .string()
    .nullable()
    .describe("ISO 8601 timestamp when the drop expires"),
  createdAt: zod.string().describe("ISO 8601 timestamp"),
  txHash: zod
    .string()
    .nullable()
    .describe("Transaction hash of the drop creation"),
});
export const ListDropsResponse = zod.array(ListDropsResponseItem);

/**
 * Register a newly deployed drop contract with the platform
 * @summary Register a new drop
 */
export const CreateDropBody = zod.object({
  contractAddress: zod.string(),
  creatorAddress: zod.string(),
  title: zod.string(),
  totalAmount: zod.string(),
  amountPerClaim: zod.string(),
  maxClaims: zod.number(),
  expiresAt: zod.string().nullish(),
  txHash: zod.string().nullish(),
});

/**
 * @summary Get a drop by contract address
 */
export const GetDropParams = zod.object({
  contractAddress: zod.coerce.string(),
});

export const GetDropResponse = zod.object({
  id: zod.number(),
  contractAddress: zod.string(),
  creatorAddress: zod.string(),
  title: zod.string(),
  totalAmount: zod
    .string()
    .describe(
      "Total USDC amount in the drop (as string to handle large numbers)",
    ),
  amountPerClaim: zod.string().describe("Amount each claimer receives in USDC"),
  maxClaims: zod.number(),
  claimedCount: zod.number(),
  isActive: zod.boolean(),
  expiresAt: zod
    .string()
    .nullable()
    .describe("ISO 8601 timestamp when the drop expires"),
  createdAt: zod.string().describe("ISO 8601 timestamp"),
  txHash: zod
    .string()
    .nullable()
    .describe("Transaction hash of the drop creation"),
});

/**
 * Returns the claim history for a specific drop
 * @summary List claims for a drop
 */
export const ListDropClaimsParams = zod.object({
  contractAddress: zod.coerce.string(),
});

export const ListDropClaimsResponseItem = zod.object({
  id: zod.number(),
  dropId: zod.number(),
  claimerAddress: zod.string(),
  amount: zod.string(),
  txHash: zod.string().nullable(),
  claimedAt: zod.string().describe("ISO 8601 timestamp"),
});
export const ListDropClaimsResponse = zod.array(ListDropClaimsResponseItem);

/**
 * Record that a wallet claimed from a drop
 * @summary Record a new claim
 */
export const RecordClaimParams = zod.object({
  contractAddress: zod.coerce.string(),
});

export const RecordClaimBody = zod.object({
  claimerAddress: zod.string(),
  amount: zod.string(),
  txHash: zod.string().nullish(),
});

/**
 * Returns recent claims across all drops for the live activity feed
 * @summary Get recent activity across all drops
 */
export const GetRecentActivityResponseItem = zod.object({
  claimerAddress: zod.string(),
  amount: zod.string(),
  dropTitle: zod.string(),
  contractAddress: zod.string(),
  claimedAt: zod.string(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);

/**
 * Returns aggregate stats - total USDC dropped, total claims, active drops count
 * @summary Get global drops summary stats
 */
export const GetDropsSummaryResponse = zod.object({
  totalDropped: zod.string().describe("Total USDC ever dropped"),
  totalClaims: zod.number(),
  activeDropsCount: zod.number(),
  totalDropsCount: zod.number(),
});