# lib/api-zod/src/generated/types/claim.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DropParty API - Viral onchain USDC giveaway platform on Arc Network
 * OpenAPI spec version: 0.1.0
 */

export interface Claim {
  id: number;
  dropId: number;
  claimerAddress: string;
  amount: string;
  /** @nullable */
  txHash: string | null;
  /** ISO 8601 timestamp */
  claimedAt: string;
}