# lib/api-zod/src/generated/types/createDropBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DropParty API - Viral onchain USDC giveaway platform on Arc Network
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDropBody {
  contractAddress: string;
  creatorAddress: string;
  title: string;
  totalAmount: string;
  amountPerClaim: string;
  maxClaims: number;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  txHash?: string | null;
}