# lib/api-zod/src/generated/types/drop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DropParty API - Viral onchain USDC giveaway platform on Arc Network
 * OpenAPI spec version: 0.1.0
 */

export interface Drop {
  id: number;
  contractAddress: string;
  creatorAddress: string;
  title: string;
  /** Total USDC amount in the drop (as string to handle large numbers) */
  totalAmount: string;
  /** Amount each claimer receives in USDC */
  amountPerClaim: string;
  maxClaims: number;
  claimedCount: number;
  isActive: boolean;
  /**
   * ISO 8601 timestamp when the drop expires
   * @nullable
   */
  expiresAt: string | null;
  /** ISO 8601 timestamp */
  createdAt: string;
  /**
   * Transaction hash of the drop creation
   * @nullable
   */
  txHash: string | null;
}