# lib/api-zod/src/generated/types/dropsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DropParty API - Viral onchain USDC giveaway platform on Arc Network
 * OpenAPI spec version: 0.1.0
 */

export interface DropsSummary {
  /** Total USDC ever dropped */
  totalDropped: string;
  totalClaims: number;
  activeDropsCount: number;
  totalDropsCount: number;
}